test: added a test and moved the end2end-testfile to the right directory #263
