+++ package.json
@@ -7,7 +7,7 @@
 		"build": "vite build",
 		"preview": "vite preview",
 		"test": "vitest",
-		"playwright-test": "playwright test tests/integration"
+		"playwright-test": "playwright test tests/end2end"
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.51.1",
